+++ src/main/java/dojo/supermarket/model/ProductQuantity.java
@@ -1,5 +1,6 @@
 package dojo.supermarket.model;
 
+import java.math.BigDecimal;
 
-public record ProductQuantity(Product product, double quantity) {
+public record ProductQuantity(Product product, BigDecimal quantity) {
 }
